introduce HazelcastJetConfigMissingCondition which checks if Jet is present on the classpath
